make some vital corrections. we should be ready to move to another step again
