+++ middleware.ts
@@ -5,14 +5,12 @@ import { authMiddleware } from '@clerk/nextjs';
 // See https://clerk.com/docs/references/nextjs/auth-middleware for more information about configuring your middleware
 export default authMiddleware({
 	publicRoutes: [
-		// '/',
+		'/',
 		'/api/:path*',
-		// '/services',
-		// '/faqs',
-		// '/contact-us',
-		// '/about-us',
-		// '/dashboard',
-		'/:path*',
+		'/services',
+		'/faqs',
+		'/contact-us',
+		'/about-us',
 	],
 });
